Move merge
