+++ python-avd/pyavd/_eos_designs/shared_utils/filtered_tenants.py
@@ -6,11 +6,11 @@ from __future__ import annotations
 from functools import cached_property
 from typing import TYPE_CHECKING
 
+from ..._utils.merge import merge
 from ...j2filters.convert_dicts import convert_dicts
 from ...j2filters.natural_sort import natural_sort
 from ...vendor.errors import AristaAvdError, AristaAvdMissingVariableError
 from ...vendor.j2.filter.range_expand import range_expand
-from ...vendor.merge import merge
 from ...vendor.utils import default, get, get_item, unique
 
 if TYPE_CHECKING:
