+++ python-avd/pyavd/_eos_designs/shared_utils/l3_interfaces.py
@@ -6,7 +6,7 @@ from __future__ import annotations
 from functools import cached_property
 from typing import TYPE_CHECKING
 
-from ...vendor.merge import merge
+from ..._utils.merge import merge
 from ...vendor.utils import get, get_item
 
 if TYPE_CHECKING:
