+++ python-avd/pyavd/_eos_designs/shared_utils/switch_data.py
@@ -6,8 +6,8 @@ from __future__ import annotations
 from functools import cached_property
 from typing import TYPE_CHECKING
 
+from ..._utils.merge import merge
 from ...j2filters.convert_dicts import convert_dicts
-from ...vendor.merge import merge
 from ...vendor.utils import get
 
 if TYPE_CHECKING:
