+++ python-avd/pyavd/_eos_designs/shared_utils/utils.py
@@ -6,8 +6,8 @@ from __future__ import annotations
 from functools import lru_cache
 from typing import TYPE_CHECKING
 
+from ..._utils.merge import merge
 from ...vendor.errors import AristaAvdError
-from ...vendor.merge import merge
 from ...vendor.utils import get, get_item, template_var
 
 if TYPE_CHECKING:
