+++ python-avd/pyavd/_eos_designs/structured_config/connected_endpoints/monitor_sessions.py
@@ -7,8 +7,8 @@ import re
 from functools import cached_property
 from typing import TYPE_CHECKING
 
+from ...._utils.merge import merge
 from ....vendor.j2.filter.range_expand import range_expand
-from ....vendor.merge import merge
 from ....vendor.strip_empties import strip_null_from_data
 from ....vendor.utils import append_if_not_duplicate, get, groupby
 from .utils import UtilsMixin
