+++ python-avd/pyavd/_eos_designs/structured_config/core_interfaces_and_l3_edge/utils.py
@@ -10,9 +10,9 @@ from ipaddress import ip_network
 from itertools import islice
 from typing import TYPE_CHECKING
 
+from ...._utils.merge import merge
 from ....j2filters.convert_dicts import convert_dicts
 from ....vendor.errors import AristaAvdMissingVariableError
-from ....vendor.merge import merge
 from ....vendor.utils import default, get, get_item
 
 if TYPE_CHECKING:
