+++ python-avd/tests/pyavd/utils/merge/test_merge.py
@@ -10,8 +10,7 @@ import os
 import pytest
 import yaml
 from pyavd._schema.avdschema import AvdSchema
-
-from ansible_collections.arista.avd.plugins.plugin_utils.merge import merge
+from pyavd._utils.merge import merge
 
 script_dir = os.path.dirname(__file__)
 with open(f"{script_dir}/access_lists.schema.yml", "r", encoding="utf-8") as schema_file:
